Added custom GPUJPEG_ASSERT
Added unconditional GPUJPEG_ASSERT to replace standard assert. The
problem with the assert() macro is that cmake for release target defines
NDEBUG resulting in the macro to be eliminated. But it is used to catch
logical errors. One alternative is abort() but it doesn't print anything
(neither of file/line/func) so better to define own assert that imitates
the built-in.