moved buffer clearing from preprocessor to init
Do not clear the comp buffer every time preprocessor is run - device
memory set causes implicit synchronization and there is perhaps small
overhead.

The oneshot clear is sufficient here - every time the input size is
changed, this code will be run but the image margins remain zeroed when
image sizes are the same.