README.md: measured perf with current code + HW
Measured also for the GTX 580 for reference (original results were on
GTX 580 only). The new numbers are significantly better (3 times) -
there may be some improvements in code (the original version was
measured in 2011, inbetween M. Jirman improved the performance). There
is also no reference which material were the original results measured
on, which can greatly influence the results.