CMake: set BUILD_OPENGL=OFF by default
Perhaps not much often used by users and it adds dependency on OpenGL,
which could be undesirable in some situations.

CI Windows build still enables the OpenGL interoperability explicitly
(perhaps not a big deal since the GLEW/GLFW libraries are statically
linked and OpenGL lib is always present in MSW).