decoder: autodetect pixfmt U8 for grayscale
This removes following error when trying to decode
a grayscale picture with output set just with
`gpujpeg_decoder_output_set_default(&decoder_output)`:
```
[GPUJPEG] [Error] Decoding single component JPEG allowed only to single component output format!
````

Note that the format is selected by adjust_format() in gpujpeg_reader
and it is grayscale (U8) for single component JPEG and RGB otherwise
(with the exception of 4 component images, in which case RGBA is decoded).
Note that color space is not deduced, still always set to GPUJPEG_RGB by
gpujpeg_image_set_default_parameter().

\+ added the symbolic name GPUJPEG_PIXFMT_AUTODETECT (+ reformat the
stuff)