encoder: allow encoding RGB to grayscale
Allow running preprocessor for 4:0:0 subsampling + added the subsampling
as a macro. It can be used to set the encode of an image monochromatic on input
explicitly as well - although not necessary, for single image, it may be needed
if the user selected the subsampling explicitly to a different format for some
of previous images.

now works:
```
gpujpegtool -e -s 1920x1080 -f 422-u8-p1020 <src>/colors/camera_bt709_422.yuv \
    -S400 camera_gray.jpg
```