gpujpegtool,lib: move statistics printout directly to the lib
Tf gpujpeg_parameters.perf_stats is set, print the measurements with
the library.

This was moved from gpujpegtool and seem to be nicer solution than
fiddling with the struct gpujpeg_duration_stats. This can be later
removed from API.

The verbosity level is the same as it was when in gpujpegtool - with
normal verbosity prints the aggregate duration per image, in verbose
per-component (eg. Huffman) statistics. This may change in future.