gpujpeg_reader_read_sos: check buf overflow early
Check the buffer overflow right at the beginning. Not needed to do
it later in the loop. Moreover, there are 3 bytes read after the
per-component reads which were not checked.

Make the length and comp_count const to make clear that those wont change
(length is sometimes decreased when parsing some markers to contain the
current remaining length).