encoder: add convenience input sett. alternatives
Add function returning gpujpeg_encoder_input structure for convenience,
mostly to make tools like clang-tidy happy that the struct is initialized
when defined. It is much better than tempting user to use something like
(`= { 0 }`) to suppress the warning. Also this can be used in C++
brace-or-equal or class initializer section.