gpujpeg_preprocessor_decode_no_transform: fix cond
When selecting dummy/memcpy decode, the incompatible input/output CS
was  considered only for ch_count=3, which is obviously insufficient.
This would perhaps allow decoding from YUVA 4:4:4:4 to RGBA without
color-space conversion.

(The idea of checking the CS difference only for 3+ channels is that if
the image contains 1 channel, we do not transcode it anyhow when in is
eg. BT.601 and out BT.709, similarly for "RGB" grayscale).