gpujpeg_image_save_to_file: (const char *) fname
gpujpeg_image_save_to_file() - take (const char *) for filename instead
of (char *).

This reverts back the change in the function prototype by the commit
de0f3a8a (2025-02-28). The (const char *) still can be modified as per
the change but it is rather a special case that is docuemented in doxy.

refer to GH-98