MSVC compat - static not supported in declaration
The C99 static in array declarator in function array list (`foo(int
a[static 3])`) is not supported by MSVC, which doesn't implement C99.

Suprisingly, [1] states: "When you specify /std:c11 or /std:c17, MSVC
supports all the features of C11 and C17 required by the standards." but
this doesn't seem to be correct for this feature (MSVS 2022).

[1]: https://learn.microsoft.com/en-us/cpp/build/reference/std-specify-language-standard-version?view=msvc-170