get_image_info: remove custom pixfmt deduction
rely rather on detection GPUJPEG_PIXFMT_AUTODETECT

This will slightly modify the behavior, eg. for non-subsmapled interleaved
image now 444-u8-p012 is returned, not 444-u8-p0p1p2 (in other words,
the planar variant is not selected based on if interleaving is used;
but this cannot be seen as a feature - the interleaved blocks and packed
pixel format are just similar concepts; the interleaving parameter is
available directly).