readded Exif custom tag writer
Removed by the commit 1510ad21 because it was originally thought that
this will be incompatible with the newly added generic metadata handling.
But (at least now) thouse 2 ways are not incompatible so keep the Exif
custom tag writer for now.

I do not won't to extend the Exif stuff much in future because it
doesn't seem that there is much useful stuff in the Exif metadata so
higher complexity may not hurt.

+ add Exif to README.md